// apps/web/packages/shared/src/_api/api.ts
export interface ChangePasswordDto {
    /**
     * 
     * @type {string}
     * @memberof ChangePasswordDto
     */
    oldPassword: string;
    /**
     * 
     * @type {string}
     * @memberof ChangePasswordDto
     */
    newPassword: string;
}
// ...
export interface ForgotPasswordDto {
    /**
     * 
     * @type {string}
     * @memberof ForgotPasswordDto
     */
    email: string;
}
// ...
export interface RefreshTokenDto {
    /**
     * 
     * @type {string}
     * @memberof RefreshTokenDto
     */
    refreshToken: string;
}
// ...
export interface SetUserPasswordDto {
    /**
     * 
     * @type {string}
     * @memberof SetUserPasswordDto
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof SetUserPasswordDto
     */
    token: string;
    /**
     * 
     * @type {string}
     * @memberof SetUserPasswordDto
     */
    password: string;
}
/**
 * 
 * @export
 * @interface SignInDto
 */
export interface SignInDto {
    /**
     * 
     * @type {string}
     * @memberof SignInDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof SignInDto
     */
    password: string;
}
// ...
export interface SignUpDto {
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    phone: string;
    /**
     * 
     * @type {string}
     * @memberof SignUpDto
     */
    password: string;
}